fixed 3.7

+++ contracts/src/instances/DsrHyperdrive.sol
@@ -185,7 +185,7 @@ contract DsrHyperdrive is Hyperdrive {
 
     /// @notice Taken from https://github.com/makerdao/dss/blob/master/src/pot.sol#L85
     /// @return z
-    function _rpow(uint x, uint n, uint base) internal pure returns (uint z) {
+    function _rpow(uint256 x, uint256 n, uint256 base) internal pure returns (uint z) {
         assembly ("memory-safe") {
             switch x
             case 0 {
